Update mempool.cpp
Changes to fix:

The success variable in mempool_sync is shared between multiple threads without proper synchronization. Since multiple threads can potentially write to this variable at the same time, the value of success may not be consistent or accurate. One way to fix this is to use an atomic variable or a mutex to synchronize access to success.

In mempool_sync, if sending transactions to any of the selected neighbors fails, the code puts all the transactions back in the toSend queue. This means that if a single transaction fails to send, all the other transactions that were queued up to be sent will also be delayed. It may be better to handle each transaction separately and only put back the ones that failed to send.

The addTransaction function checks whether a transaction is already in the mempool by checking the transactionQueue set using find. However, this does not take into account the fact that a transaction may have the same hash as another transaction but be different (due to different input or output values, for example). A better way to check for duplicate transactions would be to compare the hashes directly.

The addTransaction function calculates the total transaction amount (including the fee) and checks if it exceeds the wallet's balance in the mempool. However, this calculation does not take into account any transactions that are already in the mempool but have not yet been included in a block. It may be better to keep track of the total outgoing amount for each wallet separately, including transactions in the mempool as well as confirmed transactions.